Revert "Feat/481 api health 체크 및 외부 api health check"
