+++ src/main/java/core/global/config/SmokeConfig.java
@@ -1,6 +1,6 @@
 package core.global.config;
 
-import core.global.smoke.utils.SmokeProperties;
+import core.global.smoke.SmokeProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
