+++ src/main/java/core/global/controller/AdminMonitoringController.java
@@ -1,4 +1,4 @@
-package core.global.admin.controller;
+package core.global.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,10 +6,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequestMapping("/admin")
-public class AdminViewController {
+public class AdminMonitoringController {
 
-    @GetMapping("/login")
-    public String adminLoginPage() {
-        return "admin/admin-login";
+    @GetMapping("/monitoring")
+    public String monitoringPage() {
+        return "admin/monitoring";
     }
 }
