+++ src/main/java/core/global/controller/AdminPostController.java
@@ -1,9 +1,10 @@
-package core.global.admin.controller;
+package core.global.controller;
 
 import core.domain.post.service.PostService;
 import core.domain.user.entity.User;
 import core.domain.user.repository.UserRepository;
 import core.global.config.CustomUserDetails;
+import core.global.enums.BoardCategory;
 import core.global.enums.errorcode.UserErrorCode;
 import core.global.exception.BusinessException;
 import lombok.RequiredArgsConstructor;
@@ -15,6 +16,8 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.List;
+
+import static core.global.enums.errorcode.UserErrorCode.USER_NOT_FOUND;
 
 @Slf4j
 @Controller
