+++ src/main/java/core/global/controller/AdminViewController.java
@@ -1,4 +1,4 @@
-package core.global.admin.controller;
+package core.global.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,10 +6,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequestMapping("/admin")
-public class AdminMonitoringController {
+public class AdminViewController {
 
-    @GetMapping("/monitoring")
-    public String monitoringPage() {
-        return "admin/monitoring";
+    @GetMapping("/login")
+    public String adminLoginPage() {
+        return "admin/admin-login";
     }
 }
