+++ src/main/java/core/global/controller/ChatAdminViewController.java
@@ -1,4 +1,4 @@
-package core.global.admin.controller;
+package core.global.controller;
 
 import core.domain.chat.dto.*;
 import core.domain.chat.entity.ChatRoom;
