+++ src/main/java/core/global/smoke/SmokeController.java
@@ -1,8 +1,5 @@
-package core.global.smoke.controller;
+package core.global.smoke;
 
-import core.global.smoke.runner.SmokeGateRunner;
-import core.global.smoke.utils.SmokeProperties;
-import core.global.smoke.dto.SmokeResult;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -24,8 +21,9 @@ public class SmokeController {
             return ResponseEntity.status(404).build();
         }
 
-        SmokeResult result = runner.run(mode);
-        return result.success()
+        // 지금은 gate만
+        SmokeResult result = runner.runGate();
+        return result.ok()
                 ? ResponseEntity.ok(result)
                 : ResponseEntity.status(500).body(result);
     }
