+++ src/main/java/core/global/smoke/SmokeResult.java
@@ -1,0 +1,14 @@
+package core.global.smoke;
+import java.util.List;
+
+public record SmokeResult(
+        String mode,
+        boolean ok,
+        int total,
+        int passed,
+        int failed,
+        long durationMs,
+        List<SmokeItem> items
+) {}
+
+
