+++ src/main/resources/application-dev.yml
@@ -224,146 +224,23 @@ smoke:
   enabled: true
   token: ${SMOKE_TOKEN:}
   base-url: ${SMOKE_BASE_URL:http://localhost:8080}
-  admin:
-    email: ${SMOKE_ADMIN_EMAIL}
-    password: ${SMOKE_ADMIN_PW}
-    login-path: ${SMOKE_ADMIN_LOGIN_URL}
   gate:
     - name: health
-      type: "HEALTH"
       method: GET
       path: /actuator/health
       expected-status: 200
 
+
   full:
-    # --- [2] 정적/기본 정보 조회 (네트워크 및 캐시 확인) ---
-    - name: "App-Version-Check"
-      type: "INTERNAL"
-      method: "GET"
-      path: "/api/v1/app/version"
-      expectedStatus: 200
-
-    - name: "My-Page-Support-Links"
-      type: "INTERNAL"
-      method: "GET"
-      path: "/api/v1/mypage/support-links"
-      expectedStatus: 200
-
-    # --- [3] 주요 데이터 조회 (DB 연결성 확인) ---
-    - name: "Get-User-Profile"
-      type: "INTERNAL"
-      method: "GET"
-      path: "/api/v1/members/infos"
-      expectedStatus: 200
-
-    - name: "Get-Follow-Lists"
-      type: "INTERNAL"
-      method: "GET"
-      path: "/api/v1/mypage/follows"
-      expectedStatus: 200
-
-    - name: "Get-Post-Lists"
-      type: "INTERNAL"
-      method: "GET"
-      path: "/api/v1/boards/1/posts" # 존재하는 ID 가정
-      expectedStatus: 200
-
-    - name: "Get-Post-Comments-Sample"
-      type: "INTERNAL"
-      method: "GET"
-      path: "/api/v1/posts/1/comments" # 존재하는 ID 가정
-      expectedStatus: 200
-
-    - name: "Get-My-Posts"
-      type: "INTERNAL"
-      method: "GET"
-      path: "/api/v1/boards/my/posts?size=20"
-      expectedStatus: 200
-
-    - name: "Get-My-Bookmarks"
-      type: "INTERNAL"
-      method: "GET"
-      path: "/api/v1/boards/my/bookmarks?size=20"
-      expectedStatus: 200
-
-    - name: "Get-Search-Posts"
-      type: "INTERNAL"
-      method: "GET"
-      path: "/api/v1/search/1/posts"
-      expectedStatus: 200
-
-    - name: "Get-Chat-Rooms-My-List"
-      type: "INTERNAL"
-      method: "GET"
-      path: "/api/v1/chat/rooms"
-      expectedStatus: 200
-
-    - name: "Friend-Recommend-Logic"
-      type: "INTERNAL"
-      method: "GET"
-      path: "/api/v1/commend/content-based"
-      expectedStatus: 200
-
-    # --- [5] 관리자/내부 도구 확인 ---
-    - name: "Admin-User-Basic-Info"
-      type: "ADMIN"
-      method: "GET"
-      path: "/api/admin/users/1/basic"
-      expectedStatus: 200
-
-    # --- 외부 의존성 확인 ---
-    # --- [1] 인프라 생존 확인 (Actuator) ---
-    - name: "Infra-Redis-Health"
-      type: "INTERNAL"
-      method: "GET"
-      path: "/actuator/health/redis"
-      expectedStatus: 200
-
-    - name: "Infra-Naver-Mail-SMTP"
-      type: "INTERNAL"
-      method: "GET"
-      path: "/actuator/health/mail"
-      expectedStatus: 200
-
-    # --- [2] NCP (Naver Cloud Platform) 의존성 ---
-    - name: "Ext-NCP-Object-Storage"
-      type: "EXTERNAL"
-      method: "GET"
-      path: "${ncp.s3.end-point}" # S3Config의 endpoint 주소
-      expectedStatus: 200 # 인증 실패 시 403이 올 수 있으나 통로 확인용
-
-    - name: "Ext-NCP-Maps-API"
-      type: "EXTERNAL"
-      method: "GET"
-      path: "https://naveropenapi.apigw.ntruss.com/map-geocode/v2/getCoord"
-      expectedStatus: 401 # API Key 없이 찔렀을 때 401이면 NCP 연결 성공
-
-    # --- [3] 소셜 인증 및 푸시 (Auth & Firebase) ---
-    - name: "Ext-Apple-Auth-Keys"
-      type: "EXTERNAL"
-      method: "GET"
-      path: "https://appleid.apple.com/auth/keys"
-      expectedStatus: 200
-
-    - name: "Ext-Firebase-FCM"
-      type: "EXTERNAL"
-      method: "GET"
-      path: "https://fcm.googleapis.com"
-      expectedStatus: 200
-
-    - name: "Ext-Google-OAuth"
-      type: "EXTERNAL"
-      method: "GET"
-      path: "https://www.googleapis.com/oauth2/v3/userinfo"
-      expectedStatus: 401
-
-    # --- [4] 기타 인프라 체크 ---
-    - name: "Infra-WebSocket-Endpoint"
-      type: "INTERNAL"
-      method: "GET"
-      path: "/ws"
-      expectedStatus: 400 # WebSocket은 일반 GET으로 찌르면 400(Bad Request) 반환이 정상 생존 신호
-
+    - name: health
+      method: GET
+      path: /actuator/health
+      expected-status: 200
+    - name: api-docs
+      method: GET
+      path: /v3/api-docs
+      expected-status: 200
+    # 여기에 “전체 API”를 계속 추가
 
 openai:
   api-key: ${OPEN_AI_KEY}
